Rename `proxy_request.rs` to `proxy_get_request.rs`
