+++ http-server/src/middleware/mod.rs
@@ -1,4 +1,4 @@
 //! Various middleware implementations for RPC specific purposes.
 
 /// Proxy `GET /path` to internal RPC methods.
-pub mod proxy_request;
+pub mod proxy_get_request;
