+++ http-server/src/middleware/proxy_get_request.rs
